refactor: Move experience and skills into subfolders
